requirements.txt
# --- Web server ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
psycopg[binary,pool]==3.2.10

# --- Environment vars ---
python-dotenv==1.0.1

# --- LangChain + Google Generative AI ---
# Use langchain-core only to avoid heavy optional deps like NumPy
langchain-core==0.2.36
langchain-google-genai==1.0.7
google-generativeai==0.7.2

# --- Utils ---
pydantic==2.9.2
pydantic-settings==2.6.0
